+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.9.0 (Upcoming)
 ### Enhancements
 * NWBZarrIO load_namespaces=True by default. @mavaylon1 [#204](https://github.com/hdmf-dev/hdmf-zarr/pull/204)
+* Added test for opening file with consolidated metadata from DANDI. @mavaylon1 [#206](https://github.com/hdmf-dev/hdmf-zarr/pull/206)
 
 ## 0.8.0 (June 4, 2024)
 ### Bug Fixes